1.增加cli函数,支持新功能   /*命令名称(cmd name)    参数数目(para num)     控制函数(cmd func)     帮助信息(help)*/   {"ls",                  0,                    filelistHandle,        "显示文件目录结构"},   {"mkdir",               1,                    mkdirHandle,           "创建文件夹"},   {"rm",                  1,                    rmHandle,              "删除文件或文件夹"},   {"cat",                 1,                    catHandle,             "显示文件内容"}

+++ kelvinlawson-atomthreads-f29a33f/hardwaredriver/spiflash.c
@@ -251,7 +251,7 @@ void spiFlashTest(void)
   
   FlashID = spiFlashReadID();
 
-  printf("\r\n 芯片的ID是 0x%X", FlashID);
+  printf("\r\n 芯片的ID是 0x%X", (int)FlashID);
 
   if (FlashID == 0xEF4015)  
   {
